// src/types/ICertificate.ts
export interface ICertificate {
    id: number
    uuid: string
    course_enrollment_id: number
    certificate_no: string
    certificate_type: string
    student_id: number
    created_at: string
    updated_at: string
    title: string
    course_id: number
    template_path: string
    name: string
    slug: string
    logo: string
    certificate_type_label: string
  }
  